Fixes and Improvements
Fixes and improvements

+++ TraceabilityDriverService/Controllers/TradingPartnerController.cs
@@ -30,16 +30,22 @@ namespace TraceabilityDriverService.Controllers
             _configuration = configuration;
         }
 
+        /// <summary>
+        /// HTTP request to return a Trading Partner of an Account by ID.
+        /// </summary>
+        /// <param name="accountID"></param>
+        /// <param name="tradingPartnerID"></param>
+        /// <returns></returns>
         [HttpGet]
         [Route("{accountID}/{tradingPartnerID}")]
-        public async Task<ITEDriverTradingPartner> Get(long accountID, long tradingPartnerID)
+        public async Task<ActionResult<ITEDriverTradingPartner>> Get(long accountID, long tradingPartnerID)
         {
             try
             {
                 using (ITEDriverDB driverDB = _configuration.GetDB())
                 {
                     ITEDriverTradingPartner tp = await driverDB.LoadTradingPartnerAsync(accountID, tradingPartnerID);
-                    return tp;
+                    return new ActionResult<ITEDriverTradingPartner>(tp);
                 }
             }
             catch (Exception Ex)
@@ -48,13 +54,29 @@ namespace TraceabilityDriverService.Controllers
                 throw;
             }
         }
-
+        /// <summary>
+        /// HTTP request to save a Trading Partner to an Account.
+        /// </summary>
+        /// <param name="accountID"></param>
+        /// <param name="pglnStr"></param>
+        /// <returns></returns>
         [HttpPost]
         [Route("{accountID}/{pglnStr}")]
-        public async Task<IActionResult> Post(long accountID, string pglnStr)
+        public async Task<ActionResult<ITEDriverTradingPartner>> Post(long accountID, string pglnStr)
         {
             try
             {
+                // validation
+                if(accountID == 0)
+                {
+                    return new BadRequestObjectResult("Account ID is not set");
+                }
+                if (string.IsNullOrEmpty(pglnStr))
+                {
+                    return new BadRequestObjectResult("PGLN is null or empty");
+                }
+
+
                 IPGLN pgln = IdentifierFactory.ParsePGLN(pglnStr);
 
                 // we need to download trading partner from the directory
@@ -73,7 +95,7 @@ namespace TraceabilityDriverService.Controllers
                         tradingPartner.AccountID = accountID;
                         await driverDB.SaveTradingPartnerAsync(tradingPartner);
 
-                        return new AcceptedResult("", tradingPartner);
+                        return new ActionResult<ITEDriverTradingPartner>(tradingPartner);
                     }
                 }
             }
@@ -84,17 +106,37 @@ namespace TraceabilityDriverService.Controllers
             }
         }
 
+        /// <summary>
+        /// HTTP Request to delete a Trading Partner from an Account.
+        /// </summary>
+        /// <param name="accountID"></param>
+        /// <param name="tradingPartnerID"></param>
+        /// <returns></returns>
         [HttpDelete]
         [Route("{accountID}/{tradingPartnerID}")]
-        public async Task Delete(long accountID, long tradingPartnerID)
+        public async Task<IActionResult> Delete(long accountID, long tradingPartnerID)
         {
             try
             {
+                // validation
+                if (accountID < 1)
+                {
+                    return new BadRequestObjectResult("Account ID not properly set");
+                }
+
+                if (tradingPartnerID < 1)
+                {
+                    return new BadRequestObjectResult("Trading Partner ID not properly set");
+                }
+
                 // we will delete it if it exists, otherwise do nothing
                 using (ITEDriverDB driverDB = _configuration.GetDB())
                 {
                     await driverDB.DeleteTradingPartnerAsync(accountID, tradingPartnerID);
                 }
+
+                return new OkResult();
+
             }
             catch (Exception Ex)
             {
